pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "otelme"
authors = [{name = "Ryan Scott Brown", email = "otelme@ryansb.com"}]
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.8"
keywords = ["instrumentation", "opentelemetry", "otel", "tracing", "observability"]
dependencies = [
    "opentelemetry-distro[otlp]",
    "opentelemetry-instrumentation",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-html",
]

[tool.pytest.ini_options]
norecursedirs = ".venv .git"
testpaths = "test_otelme.py"
addopts = "--cov=otelme --cov-report=html --html=./htmlcov/tests.html --strict-config --strict-markers"
[tool.black]
line-length = 120
target-version = ['py39']
